añadiendo css selector excecutor en el mismo archivo sin la n
